Revert "Move Effect into the internal algebra module."
This reverts commit 06d2fc07697be20d8a23359eaa740588af21b4f1.
